[XLA:GPU] Trigger a crash if nan_counter check was requested
The kernel could be unstable and produce NaN values.
When one want to detect such cases they need to set the flag xla_gpu_experimental_enable_nan_counter_on_thunks to true.
In this case the counter for NaN values will be triggered
and if there is any then the execution will be crashed.

PiperOrigin-RevId: 827895646